SCRUM-4: Move endpoint definitions to sports_api/endpoints package
